+++ leaderboard1.html
@@ -1,0 +1,1 @@
+<html><body><div align="center" style="margin-top:4rem"><h3>leaderboard1</h3><table border=1><tr><th style="padding:0.5rem">Rank</th><th style="padding:0.5rem">Name</th></tr><tr><td style="padding:0.5rem">1</td><td style="padding:0.5rem">bill</td></tr><tr><td style="padding:0.5rem">2</td><td style="padding:0.5rem">bob</td></tr><tr><td style="padding:0.5rem">3</td><td style="padding:0.5rem">mike</td></tr><tr><td style="padding:0.5rem">4</td><td style="padding:0.5rem">fred</td></tr><tr><td style="padding:0.5rem">5</td><td style="padding:0.5rem">sandeep</td></tr><tr><td style="padding:0.5rem">6</td><td style="padding:0.5rem">dick</td></tr><tr><td style="padding:0.5rem">7</td><td style="padding:0.5rem">reggie</td></tr><tr><td style="padding:0.5rem">8</td><td style="padding:0.5rem">sandeeps</td></tr><tr><td style="padding:0.5rem">9</td><td style="padding:0.5rem">reg</td></tr><tr><td style="padding:0.5rem">10</td><td style="padding:0.5rem">geg</td></tr><tr><td style="padding:0.5rem">11</td><td style="padding:0.5rem">red</td></tr><tr><td style="padding:0.5rem">12</td><td style="padding:0.5rem">ss</td></tr><tr><td style="padding:0.5rem">13</td><td style="padding:0.5rem">bobs</td></tr></table></div></body></html>
